Huge refactor - HUGE
